input/touchscreen: Change the type of touch_upperhalf_s::nums from uint32_t to uint8_t
to align with the type of touch_register

+++ drivers/input/touchscreen_upper.c
@@ -56,7 +56,7 @@ struct touch_openpriv_s
 
 struct touch_upperhalf_s
 {
-  uint32_t         nums;               /* Number of touch point structure */
+  uint8_t          nums;               /* Number of touch point structure */
   sem_t            exclsem;            /* Manages exclusive access to this structure */
   struct list_node head;               /* Opened file buffer chain header node */
   FAR struct touch_lowerhalf_s *lower; /* A pointer of lower half instance */
